+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## [3.0.0](https://github.com/danielspk/TornadoHttp/releases/tag/v3.0.0) (201x-xx-xx)
 
 * *Core: PHP 7.2 version*
+* *Core: add handler context*
 * *Core: remove auto default response*
+* *Core: remove globalAttributes*
 * *Core: set initial response*
 
 ## [2.1.3](https://github.com/danielspk/TornadoHttp/releases/tag/v2.1.3) (2018-09-20)